cli/v2/merge/requirements.txt
typer
rich
Pillow
mutagen
rich